backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ytmusicapi==1.8.2
yt-dlp==2024.8.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.9.1
